Create README.md
This is a simple guess the number game that allows a player to choose a random number between 1 and 100, then challenge the player to guess the number in 10 turns. After each turn, the player is told if they are right or wrong and if they are wrong, whether the guess was too high or too low. The game will end once the player guesses correctly or once they run out of turns. Once the game ends, the player is given an option to start playing again.